+++ db/column_family.cc
@@ -990,12 +990,14 @@ const int ColumnFamilyData::kCompactToBaseLevel = -2;
 Compaction* ColumnFamilyData::CompactRange(
     const MutableCFOptions& mutable_cf_options, int input_level,
     int output_level, uint32_t output_path_id, uint32_t max_subcompactions,
+    BottommostLevelCompaction bottommost_level_compaction,
     const InternalKey* begin, const InternalKey* end,
     InternalKey** compaction_end, bool* conflict,
     uint64_t max_file_num_to_ignore) {
   auto* result = compaction_picker_->CompactRange(
       GetName(), mutable_cf_options, current_->storage_info(), input_level,
-      output_level, output_path_id, max_subcompactions, begin, end,
+      output_level, output_path_id, max_subcompactions,
+      bottommost_level_compaction, begin, end,
       compaction_end, conflict, max_file_num_to_ignore);
   if (result != nullptr) {
     result->SetInputVersion(current_);