+++ db/compaction_picker.h
@@ -59,6 +59,7 @@ class CompactionPicker {
       const std::string& cf_name, const MutableCFOptions& mutable_cf_options,
       VersionStorageInfo* vstorage, int input_level, int output_level,
       uint32_t output_path_id, uint32_t max_subcompactions,
+      BottommostLevelCompaction bottommost_level_compaction,
       const InternalKey* begin, const InternalKey* end,
       InternalKey** compaction_end, bool* manual_conflict,
       uint64_t max_file_num_to_ignore);
@@ -258,6 +259,7 @@ class NullCompactionPicker : public CompactionPicker {
                            int /*input_level*/, int /*output_level*/,
                            uint32_t /*output_path_id*/,
                            uint32_t /*max_subcompactions*/,
+                           BottommostLevelCompaction /*bottommost_level_compaction*/,
                            const InternalKey* /*begin*/,
                            const InternalKey* /*end*/,
                            InternalKey** /*compaction_end*/,